trying to figure out backtracking still
What I understand from backtracking is its basically a recursion call that returns to a previous state and continues from there if there is an invalidation condition met. So basically traversing through something  and then returning to my most correct (or in most conditions the only correct point) to continue to the end. Still trying to figure out the structuring needed and the for loops of multi index traversing are still a little confusing. I need to work on the planning part of the text question before jumping into the programming.