Understand BFS pretty good now
Things I learned about BFS:
It traverses the nodes in level order left to right
It has levels
It uses a while loop and not recursion
different code structure using while len(queue) != 0 to keep the loop going
uses a queue instead of a stack. We can pop things from a queue at the front or back easier than a stack.